pyproject.toml
[tool.poetry]
name = "goit_pyweb_hw_11"
version = "0.1.0"
description = ""
authors = ["Ilya Dudelzak <65915836+IlyaDudelzak@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
fastapi = "^0.115.11"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
SQLAlchemy = "^2.0.38"
psycopg2 = "^2.9.10"
pydantic = {extras = ["email"], version = "^2.10.6"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
